+++ lib/nanoc/base/compilation/item_rep_proxy.rb
@@ -59,11 +59,12 @@ module Nanoc
     # @param [String] layout_identifier The identifier of the layout to use
     #
     # @return [void]
-    def layout(layout_identifier)
+    def layout(layout_identifier, extra_filter_args={})
       set_assigns
 
       layout = layout_with_identifier(layout_identifier)
       filter_name, filter_args = @compiler.rules_collection.filter_for_layout(layout)
+      filter_args = filter_args.merge(extra_filter_args)
 
       @item_rep.layout(layout, filter_name, filter_args)
     end