+++ NEWS.md
@@ -7,6 +7,7 @@ Base:
 * Dropped the “3” suffix on nanoc3/Nanoc3
 * Turned Rake tasks into proper nanoc commands
 * Improved dependency tracking
+* Added support for locals in filters and layouts
 
 Extensions: